requirements.txt
mysql-connector-python==8.0.32
python-telegram-bot==20.0